update kernel
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@66208 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/universal/linux-6.1-nss/fs/udf/inode.c
@@ -398,7 +398,7 @@ static int udf_do_extend_file(struct inode *inode,
 			  sb->s_blocksize - 1) & ~(sb->s_blocksize - 1));
 		iinfo->i_lenExtents =
 			(iinfo->i_lenExtents + sb->s_blocksize - 1) &
-			~(sb->s_blocksize - 1);
+			~((u64)sb->s_blocksize - 1);
 	}
 
 	/* Can we merge with the previous extent? */
